feat: check constraints
